+++ src/assets/styles/style.scss
@@ -1,6 +1,23 @@
 $color-pink: #f092db;
 $color-green: #92f0a2;
 
+body {
+  display: flex;
+  flex-direction: column;
+}
+h1 {
+  text-align: center;
+  color: #7a797a;
+}
+
+textarea {
+  display: flex;
+  // width: 1022px;
+  // height: 120px;
+  border: 2px solid #d4d3d4;
+  margin: 20px auto 0;
+}
+
 .keyboard-container {
   display: flex;
   flex-direction: column;
@@ -36,8 +53,8 @@ $color-green: #92f0a2;
 			width: 48px;
 			flex-wrap: wrap;
 			min-height: 36px;
-			// display: inline-block;
-			text-transform: capitalize;
+      color: #8d8b8b;
+			// text-transform: capitalize;
 			font-weight: 600;
       position: relative;
       &[data-key="8"] {
@@ -59,11 +76,18 @@ $color-green: #92f0a2;
         bottom:  5px;
         right: 5px;
       }
+      &:hover {
+        background-color: #bebebe;
+      }
     }
+    
     .click {
-      transform: scale(1.1);
-      border-color: #f092db;
-      background-color: aqua;
+      transform: scale(1.05);
+      box-shadow: 0px 1px 3px;
+      background-color: #bebebe;
+    }
+    .lowercase {
+      text-transform: capitalize;
     }
   }
 }
